Make all_cells_lit recognize solved maps
Before this addition, our program could not recognize solved maps if we provided a KB that only contains wall, wall_num, and light predicates.

This is because we are checking for the dynamic `lit` predicate to conclude whether cells are lit or not.

Now if a cell is not `lit`, the program will fall back to legacy behavior, which is searching for lights in the cell's X and Y rays. Otherwise, the `lit` predicate is enough.